+++ src/org/lessons/java/poligoni/Quadrato.java
@@ -1,21 +1,21 @@
 package org.lessons.java.poligoni;
 
-public class Quadrato extends AbstractPoligono{
+public class Quadrato extends AbstractPoligono {
 
-    private int lato;
+    private double lato;
 
-    public Quadrato(String nome, int lato){
+    public Quadrato(String nome, double lato) {
         super(nome);
         this.lato = lato;
     }
 
     @Override
-    public int calcolaArea(){
+    public double calcolaArea() {
         return lato * lato;
     }
 
     @Override
-    public int calcolaPerimetro(){
+    public double calcolaPerimetro() {
         return lato * 4;
     }
