+++ src/org/lessons/java/poligoni/Rettangolo.java
@@ -1,25 +1,24 @@
 package org.lessons.java.poligoni;
 
-public class Rettangolo extends AbstractPoligono{
+public class Rettangolo extends AbstractPoligono {
 
-    private int base;
-    private int altezza;
+    private double base;
+    private double altezza;
 
-    public Rettangolo(String nome, int base, int altezza){
+    public Rettangolo(String nome, double base, double altezza) {
         super(nome);
         this.base = base;
         this.altezza = altezza;
     }
 
     @Override
-    public int calcolaArea(){
+    public double calcolaArea() {
         return base * altezza;
     }
 
     @Override
-    public int calcolaPerimetro(){
+    public double calcolaPerimetro() {
         return 2 * (base + altezza);
     }
-    
 
 }
